fix: mongo db indexes and expiration policies for ticket defns

+++ api/cas-server-core-api-configuration-model/src/main/java/org/apereo/cas/configuration/model/core/ticket/ProxyTicketProperties.java
@@ -25,10 +25,10 @@ public class ProxyTicketProperties implements Serializable {
     /**
      * Number of uses allowed.
      */
-    private int numberOfUses = 1;
+    private long numberOfUses = 1;
 
     /**
      * Number of seconds after which this ticket becomes invalid.
      */
-    private int timeToKillInSeconds = 10;
+    private long timeToKillInSeconds = 10;
 }
